fix(function-server) refactoring
