+++ services/rotor/build.mts
@@ -35,11 +35,46 @@ const externalModules = [...Object.keys(nativeDeps), "pg-native"];
 // esbuild's platform: "node" normally bundles these as bare require("fs") etc.,
 // but Deno requires the "node:" prefix. This plugin rewrites them to external "node:*" imports.
 const nodeBuiltins = [
-  "assert", "buffer", "child_process", "cluster", "console", "constants", "crypto",
-  "dgram", "dns", "domain", "events", "fs", "fs/promises", "http", "http2", "https",
-  "inspector", "module", "net", "os", "path", "perf_hooks", "process", "punycode",
-  "querystring", "readline", "repl", "stream", "string_decoder", "sys", "timers",
-  "tls", "tty", "url", "util", "v8", "vm", "wasi", "worker_threads", "zlib",
+  "assert",
+  "buffer",
+  "child_process",
+  "cluster",
+  "console",
+  "constants",
+  "crypto",
+  "dgram",
+  "dns",
+  "domain",
+  "events",
+  "fs",
+  "fs/promises",
+  "http",
+  "http2",
+  "https",
+  "inspector",
+  "module",
+  "net",
+  "os",
+  "path",
+  "perf_hooks",
+  "process",
+  "punycode",
+  "querystring",
+  "readline",
+  "repl",
+  "stream",
+  "string_decoder",
+  "sys",
+  "timers",
+  "tls",
+  "tty",
+  "url",
+  "util",
+  "v8",
+  "vm",
+  "wasi",
+  "worker_threads",
+  "zlib",
 ];
 
 function denoNodePrefixPlugin(): esbuild.Plugin {
@@ -99,7 +134,7 @@ esbuild
   .then(() => {
     // Deno workspace worker (ESM – runs in Web Worker sandbox with permissions: "none")
     return esbuild.build({
-      entryPoints: ["./src/workspace-worker.ts"],
+      entryPoints: ["./src/lib/workspace-worker.ts"],
       bundle: true,
       platform: "node",
       target: "es2022",
